Update findBy[social] methods + remove UC#user
UserCollection#user was clunky and bad, so removed it in favor of using more specific methods.

Also the findByFacebook/Twitter/Username/Email methods have been updated to return a single record if a single key is passed, an array if an array is passed, and false if no users are found.

Finally, updated authentication to use unsafeLogin + have a more intuitive API.